Create LFP submodule
